refactor: RacingInitService 클래스 분리
RacingService 클래스에서 경주를 위한 값들을 생성하는 메서드들을 RacingInitService 클래스로 분리시켰습니다.
