+++ src/main/java/racingcar/controller/RacingController.java
@@ -6,44 +6,39 @@ import racingcar.domain.Car;
 import racingcar.dto.CarDto;
 import racingcar.mapper.CarDtoMapper;
 import racingcar.observer.Observer;
+import racingcar.service.RacingInitService;
 import racingcar.service.RacingService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class RacingController implements Observer {
     private final RacingService racingService = new RacingService();
+    private final RacingInitService racingInitService = new RacingInitService();
     private List<String> carNames;
     private int attemptCount;
     private LinkedHashSet<Car> cars;
-    private LinkedHashSet<CarDto> carDtos;
 
     public void run() {
-        setRacing();
-        createCars();
-        addObserverToCar(cars);
+        readyRace();
         startRace();
     }
 
-    private void setRacing() {
+    private void readyRace() {
         OutputView.printCarNameInputMessage();
         carNames = racingService.carValidate(getUserInput());
 
         OutputView.printAttemptCountInputMessage();
-        attemptCount = Integer.parseInt(racingService.attemptCountValidate(getUserInput()));
-    }
+        attemptCount = Integer.parseInt(racingInitService.attemptCountValidate(getUserInput()));
 
-    private void createCars() {
-        cars = racingService.createCars(carNames);
-    }
-    private void createCarDtos() {
-        carDtos = racingService.creatCarDtos(cars);
+        createCars();
+        addObserverToCar(cars);
     }
 
     private void startRace() {
         OutputView.printGameResultMessage();
 
         racingService.moveAllCar(attemptCount, cars, OutputView::printEachRoundSeparator);
-        createCarDtos();
+        LinkedHashSet<CarDto> carDtos = racingInitService.createCarDtos(cars);
         List<String> winners = racingService.generateRaceResult(carDtos);
 
         OutputView.printWinner(winners);
@@ -51,6 +46,10 @@ public class RacingController implements Observer {
 
     private String getUserInput() {
         return InputView.getUserInput();
+    }
+
+    private void createCars() {
+        cars = racingInitService.createCars(carNames);
     }
 
     void addObserverToCar(LinkedHashSet<Car> cars) {
