+++ src/main/java/racingcar/service/RacingService.java
@@ -4,7 +4,6 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import racingcar.domain.Car;
 import racingcar.dto.CarDto;
-import racingcar.mapper.CarDtoMapper;
 import racingcar.model.InputCarNameSplitter;
 import racingcar.validator.ValidatorFactory;
 
@@ -15,22 +14,6 @@ public class RacingService {
         List<String> carNames = splitter.splitInputCarNames(validatedCarNameInput);
 
         return ValidatorFactory.forIndividualCarName().validate(carNames);
-    }
-    public String attemptCountValidate(String countInput) {
-        return ValidatorFactory.forAttemptCount().validate(countInput);
-    }
-
-    public LinkedHashSet<Car> createCars(List<String> validatedCarNames) {
-        RaceCarEnroller enroller = new RaceCarEnroller();
-        return enroller.enrollCars(validatedCarNames);
-    }
-
-    public LinkedHashSet<CarDto> creatCarDtos(LinkedHashSet<Car> cars) {
-        LinkedHashSet<CarDto> carDtos = new LinkedHashSet<>();
-        for (Car car : cars) {
-            carDtos.add(CarDtoMapper.from(car));
-        }
-        return carDtos;
     }
 
     public void moveAllCar(int attemptCount, LinkedHashSet<Car> cars, Runnable afterEachRound) {
